Add the `psComments` field
There will be groups of comments followed by codes after formatting a source file with `HIndent`, and it is like this:

```haskell
--- Comment 1
f1 :: Int
f1 = 3

--- Comment 2
f2 :: Int
f2 = 5
```

We must collect comments that are above a node, comments that are on the same line, and comments that are after the node in order. However, each node in the parsed result only contains comments after the node. `f1` in the above node has `-- Comment 2`, not `-- Comment 1` (and by the way, `hsmodAnn` has `-- Comment 1`).

I will change the code to collect all comments after parsing the code and store them in `psComments`. The `pretty` function will call `collectCommentsBefore`, `collectCommentsOnSameLine`, and `collectCommentsAfter`. These functions drain comments from `psComments`. I am sure that this way is too slow, but I could not come up with a better implementation.