Remove `relocateCommentsSameLineAfterNode`
First, all tests passed without this function. Next, I tested it with
the actual code, and the comments were placed correctly regardless of
whether the function was used. Finally, I tried deleting the
corresponding sections in the original HIndent code. However, it only
inserted an extra blank line, and everything else was fine. In the first
place, the behavior of the test and the actual program was different
from the original HIndent. When the code that failed the test (i.e., the
test program inserted an extra blank line) was executed in the
actual HIndent program, the blank line was not inserted. Therefore, I do
not believe that this function is necessary.