Define own `Extension`
So far, the enabled extensions have been stored in Cabal's `Extension`
in `Config`, but the `KnownExtension` included in `Extension` may not
have the latest extensions. Then, parsing may be performed with such
extensions disabled, resulting in parsing errors even for code that can
be compiled with GHC.

This commit defines its own `Extension` type. This type uses the
`Extension` of `ghc-lib-parser` internally, allowing the latest
extensions to be used at all times, thus solving the above problem.