Do not append a trailing newline
The comment says "Hack to work around some parser issues in
haskell-src-exts", but we don't use it anymore. And it also says "Some
pragmas need to have a newline following them in order to parse
properly", but I could not find such pragmas. Furthermore, adding a
trailing newline ignores the `force-trailing-newline` option.